Swichted to float for the datatype in ombvh5
double is useless since Coin/OpenGL only supports float and the data
from the HDF5 file is always (more or less) just passed to OpenGL with
no/less calculation

+++ openmbvcppinterface/openmbvcppinterface/arrow.h
@@ -29,7 +29,7 @@ namespace OpenMBV {
   /** A arrow with zero, one or two heads
    *
    * HDF5-Dataset: The HDF5 dataset of this object is a 2D array of
-   * double precision values. Each row represents one dataset in time.
+   * single or double precision values. Each row represents one dataset in time.
    * A row consists of the following columns in order given in
    * world frame: time,
    * "to" point x, "to" point y,
@@ -55,7 +55,7 @@ namespace OpenMBV {
       std::string pathStr;
       void createHDF5File() override;
       void openHDF5File() override;
-      H5::VectorSerie<double>* data{nullptr};
+      H5::VectorSerie<Float>* data{nullptr};
       double headDiameter{0.5};
       double headLength{0.75};
       double diameter{0.25};
@@ -78,7 +78,7 @@ namespace OpenMBV {
         if(row.size()!=8) throw std::runtime_error("the dimension does not match");
         if(!std::isnan(dynamicColor))
         {
-          std::vector<double> tmprow(8);
+          std::vector<Float> tmprow(8);
           std::copy(&row[0], &row[8], tmprow.begin());
           tmprow[7]=dynamicColor;
           data->append(tmprow);
@@ -88,7 +88,7 @@ namespace OpenMBV {
       }
 
       int getRows() override { return data?data->getRows():0; }
-      std::vector<double> getRow(int i) override { return data?data->getRow(i):std::vector<double>(8); }
+      std::vector<Float> getRow(int i) override { return data?data->getRow(i):std::vector<Float>(8); }
 
       /** Convenience; see setHeadDiameter and setHeadLength */
       void setArrowHead(double diameter, double length) {
